PisApp/requirements.txt
flask
argon2_cffi
mysqlclient
flask-sqlalchemy 
flask-login
gunicorn
Werkzeug
websockets
cryptography
pyOpenSSL
pycryptodome